evm vanchor leaves watcher #160
