+++ src/events_watcher/evm/mod.rs
@@ -1,0 +1,163 @@
+// Copyright 2022 Webb Technologies Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+use super::*;
+use crate::config;
+use std::ops;
+use std::sync::Arc;
+use std::time::Duration;
+use webb::evm::contract::protocol_solidity::{
+    FixedDepositAnchorContract, VAnchorContract,
+};
+use webb::evm::ethers::prelude::{Contract, Middleware};
+use webb::evm::ethers::types;
+
+mod anchor_leaves_watcher;
+mod anchor_watcher;
+mod vanchor_leaves_watcher;
+#[doc(hidden)]
+pub use anchor_leaves_watcher::*;
+#[doc(hidden)]
+pub use anchor_watcher::*;
+#[doc(hidden)]
+pub use vanchor_leaves_watcher::*;
+
+/// AnchorContractWrapper contains FixedDepositAnchorContract contract along with configurations for Anchor contract, and Relayer.
+#[derive(Clone, Debug)]
+pub struct AnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    pub config: config::AnchorContractConfig,
+    pub webb_config: config::WebbRelayerConfig,
+    pub contract: FixedDepositAnchorContract<M>,
+}
+
+impl<M> AnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    /// Creates a new AnchorContractOverDKGWrapper.
+    pub fn new(
+        config: config::AnchorContractConfig,
+        webb_config: config::WebbRelayerConfig,
+        client: Arc<M>,
+    ) -> Self {
+        Self {
+            contract: FixedDepositAnchorContract::new(
+                config.common.address,
+                client,
+            ),
+            config,
+            webb_config,
+        }
+    }
+}
+
+impl<M> ops::Deref for AnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    type Target = Contract<M>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.contract
+    }
+}
+
+impl<M> super::WatchableContract for AnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    fn deployed_at(&self) -> types::U64 {
+        self.config.common.deployed_at.into()
+    }
+
+    fn polling_interval(&self) -> Duration {
+        Duration::from_millis(self.config.events_watcher.polling_interval)
+    }
+
+    fn max_events_per_step(&self) -> types::U64 {
+        self.config.events_watcher.max_events_per_step.into()
+    }
+
+    fn print_progress_interval(&self) -> Duration {
+        Duration::from_millis(
+            self.config.events_watcher.print_progress_interval,
+        )
+    }
+}
+
+// VAnchorContractWrapper contains VAnchorContract contract along with configurations for Anchor contract, and Relayer.
+#[derive(Clone, Debug)]
+pub struct VAnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    pub config: config::VAnchorContractConfig,
+    pub webb_config: config::WebbRelayerConfig,
+    pub contract: VAnchorContract<M>,
+}
+
+impl<M> VAnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    /// Creates a new VAnchorContractOverDKGWrapper.
+    pub fn new(
+        config: config::VAnchorContractConfig,
+        webb_config: config::WebbRelayerConfig,
+        client: Arc<M>,
+    ) -> Self {
+        Self {
+            contract: VAnchorContract::new(config.common.address, client),
+            config,
+            webb_config,
+        }
+    }
+}
+
+impl<M> ops::Deref for VAnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    type Target = Contract<M>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.contract
+    }
+}
+
+impl<M> super::WatchableContract for VAnchorContractWrapper<M>
+where
+    M: Middleware,
+{
+    fn deployed_at(&self) -> types::U64 {
+        self.config.common.deployed_at.into()
+    }
+
+    fn polling_interval(&self) -> Duration {
+        Duration::from_millis(self.config.events_watcher.polling_interval)
+    }
+
+    fn max_events_per_step(&self) -> types::U64 {
+        self.config.events_watcher.max_events_per_step.into()
+    }
+
+    fn print_progress_interval(&self) -> Duration {
+        Duration::from_millis(
+            self.config.events_watcher.print_progress_interval,
+        )
+    }
+}
